# go-api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A high-performance documentation platform with intelligent search using full-text and semantic vector search.",
        "title": "Document Hub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/kavishankarks/document-hub",
            "email": "support@documenthub.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/pipeline/start": {
            "post": {
                "description": "Starts a new pipeline to process course curriculum with web search, normalization, chunking, and embedding",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Start a new RAG pipeline",
                "parameters": [
                    {
                        "description": "Pipeline configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.StartPipelineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.PipelineRun"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pipeline/{id}/cancel": {
            "post": {
                "description": "Cancels a running pipeline execution",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Cancel pipeline",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pipeline Run ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pipeline/{id}/results": {
            "get": {
                "description": "Retrieves the documents and chunks created by a pipeline run",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Get pipeline results",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pipeline Run ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.PipelineResultsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pipeline/{id}/status": {
            "get": {
                "description": "Retrieves the current status and progress of a pipeline run",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Get pipeline status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pipeline Run ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.PipelineStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pipelines": {
            "get": {
                "description": "Retrieves a paginated list of all pipeline runs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "List pipeline runs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of records to skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Maximum number of records to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, processing, completed, failed)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/documents/upload": {
            "post": {
                "description": "Uploads a file (PDF, Doc, Word, PPT, HTML), converts it to markdown, normalizes, chunks, embeds, and stores it.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Upload and process a document",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Document file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.Document"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_kavishankarks_document-hub_go-api_internal_models.Curriculum": {
            "type": "object",
            "required": [
                "modules",
                "title"
            ],
            "properties": {
                "modules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.CurriculumModule"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_kavishankarks_document-hub_go-api_internal_models.CurriculumModule": {
            "type": "object",
            "required": [
                "name",
                "topics"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_kavishankarks_document-hub_go-api_internal_models.Document": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "doc_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "object"
                },
                "source_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_kavishankarks_document-hub_go-api_internal_models.PipelineConfig": {
            "type": "object",
            "properties": {
                "chunk_overlap": {
                    "type": "integer"
                },
                "chunk_size": {
                    "type": "integer"
                },
                "normalize": {
                    "type": "boolean"
                },
                "search_engine": {
                    "description": "duckduckgo, brave",
                    "type": "string"
                },
                "search_results_per_topic": {
                    "type": "integer"
                },
                "web_search_enabled": {
                    "type": "boolean"
                }
            }
        },
        "github_com_kavishankarks_document-hub_go-api_internal_models.PipelineResultsResponse": {
            "type": "object",
            "properties": {
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.Document"
                    }
                },
                "pipeline_run": {
                    "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.PipelineRun"
                },
                "total_chunks": {
                    "type": "integer"
                }
            }
        },
        "github_com_kavishankarks_document-hub_go-api_internal_models.PipelineRun": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "current_stage": {
                    "description": "parse, search, normalize, chunk, embed, store",
                    "type": "string"
                },
                "curriculum_title": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "input_data": {
                    "type": "object"
                },
                "progress": {
                    "description": "0-100",
                    "type": "integer"
                },
                "status": {
                    "description": "pending, processing, completed, failed",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_kavishankarks_document-hub_go-api_internal_models.PipelineStatusResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_stage": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "progress": {
                    "type": "integer"
                },
                "stages": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_kavishankarks_document-hub_go-api_internal_models.StartPipelineRequest": {
            "type": "object",
            "required": [
                "curriculum"
            ],
            "properties": {
                "config": {
                    "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.PipelineConfig"
                },
                "curriculum": {
                    "$ref": "#/definitions/github_com_kavishankarks_document-hub_go-api_internal_models.Curriculum"
                }
            }
        },
...
    }
}